refactor: HandlerExecution 생성 책임 변경
