+++ src/main/java/core/mvc/tobe/AnnotationHandlerMapping.java
@@ -1,24 +1,48 @@
 package core.mvc.tobe;
 
+import static java.util.Arrays.asList;
+
+import java.lang.reflect.Method;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.core.LocalVariableTableParameterNameDiscoverer;
+import org.springframework.core.ParameterNameDiscoverer;
 
 import com.google.common.collect.Maps;
 
+import core.annotation.web.RequestMapping;
 import core.annotation.web.RequestMethod;
-import core.di.BeanScanner;
+import core.di.scanner.BeanScanner;
 import core.mvc.HandlerMapping;
+import core.mvc.tobe.support.ArgumentResolver;
+import core.mvc.tobe.support.HttpRequestArgumentResolver;
+import core.mvc.tobe.support.HttpResponseArgumentResolver;
+import core.mvc.tobe.support.ModelArgumentResolver;
+import core.mvc.tobe.support.PathVariableArgumentResolver;
+import core.mvc.tobe.support.RequestParamArgumentResolver;
 
 public class AnnotationHandlerMapping implements HandlerMapping {
     private static final Logger logger = LoggerFactory.getLogger(AnnotationHandlerMapping.class);
 
-    private BeanScanner beanScanner;
-
     private Map<HandlerKey, HandlerExecution> handlerExecutions = Maps.newHashMap();
+
+    private final BeanScanner beanScanner;
+
+    private static final List<ArgumentResolver> argumentResolvers = asList(
+            new HttpRequestArgumentResolver(),
+            new HttpResponseArgumentResolver(),
+            new RequestParamArgumentResolver(),
+            new PathVariableArgumentResolver(),
+            new ModelArgumentResolver());
+
+    private static final ParameterNameDiscoverer nameDiscoverer = new LocalVariableTableParameterNameDiscoverer();
 
     public AnnotationHandlerMapping(Object... basePackage) {
         beanScanner = new BeanScanner(basePackage);
@@ -26,7 +50,7 @@ public class AnnotationHandlerMapping implements HandlerMapping {
 
     public void initialize() {
         logger.info("## Initialized Annotation Handler Mapping");
-        handlerExecutions.putAll(beanScanner.getHandlerExecutions());
+        handlerExecutions.putAll(createHandlerExecutions());
     }
 
     public Object getHandler(HttpServletRequest request) {
@@ -42,7 +66,29 @@ public class AnnotationHandlerMapping implements HandlerMapping {
                 return handlerExecutions.get(handlerKey);
             }
         }
-
         return null;
+    }
+
+    private Map<HandlerKey, HandlerExecution> createHandlerExecutions() {
+        Map<HandlerKey, HandlerExecution> handlers = new HashMap<>();
+        Map<Class<?>, Object> controllers = beanScanner.getControllers();
+        for (Class<?> controller : controllers.keySet()) {
+            Object target = controllers.get(controller);
+            addHandlerExecution(handlers, target, controller.getMethods());
+        }
+
+        return handlers;
+    }
+
+    private void addHandlerExecution(Map<HandlerKey, HandlerExecution> handlers, final Object target, Method[] methods) {
+        Arrays.stream(methods)
+              .filter(method -> method.isAnnotationPresent(RequestMapping.class))
+              .forEach(method -> {
+                  RequestMapping requestMapping = method.getAnnotation(RequestMapping.class);
+                  HandlerKey handlerKey = new HandlerKey(requestMapping.value(), requestMapping.method());
+                  HandlerExecution handlerExecution = new HandlerExecution(nameDiscoverer, argumentResolvers, target, method);
+                  handlers.put(handlerKey, handlerExecution);
+                  logger.info("Add - method: {}, path: {}, HandlerExecution: {}", requestMapping.method(), requestMapping.value(), method.getName());
+              });
     }
 }
